requirements.txt
anyio==4.1.0
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
cssselect==1.2.0
fake-useragent==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.8.0
loadenv==0.1.1
lxml==4.9.3
outcome==1.3.0.post0
parse==1.20.0
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
urllib3==1.26.18
w3lib==2.1.2
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.17.0
